# contract/openapi.yaml
openapi: 3.1.0
info:
  title: КОСМОБЕРЕГ API
  version: 1.0.0
  description: |
    Контракт между фронтендом и бэкендом платформы КОСМОБЕРЕГ.

    ## Как этим пользоваться

    Это **единственный источник истины** об интерфейсе между фронтом и бэком.
    Фронт кодируется против контракта (с моками, пока бэка нет), бэк реализует
    контракт. Стыковка = смена базового URL.

    Изменение контракта — только по согласованию обеих сторон. Правило простое:
    **сначала правим contract/openapi.yaml, потом код.**

    ## Ключевое решение: локализация на клиенте

    API **никогда не возвращает переведённый текст**. Возвращаются доменные коды
    (`status: "problem"`, `role: "observer"`) и ключи локализации
    (`name_key: "seg.ustie.name"`). Клиент переводит их через `i18n.js`.

    Причины:
    - фронт уже несёт словарь на 3 языка (ru/en/zh) — дублировать его на бэке не нужно;
    - смена языка не требует запроса к серверу;
    - добавление 4-го языка не затрагивает бэкенд вообще.

    Исключение — **пользовательский контент** (описания в отчётах, имена
    пользователей): возвращается как есть, без перевода.

    ## Что вычисляется только на сервере

    Клиент не может это подделать, поэтому такие значения приходят с бэка и
    никогда не принимаются от фронта:

    - `attention_index` и факторы `s`, `c`, `t`, `a`;
    - переход роли `student → observer` (только после завершения курса);
    - достижение консенсуса по разметке;
...
    - смена статуса участка.
servers:
...
security:
  - bearerAuth: []
paths:
...
  /segments:
    get:
      tags: [segments]
      summary: Участки береговой линии
      description: |
        Основной источник данных для карты и для кабинета ООПТ.
        По умолчанию отсортированы по убыванию индекса внимания.
      parameters:
        - name: oopt_id
          in: query
          schema: { type: string }
          description: Фильтр по территории
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/SegmentStatus' }
          style: form
          explode: false
        - name: sort
          in: query
          schema:
            type: string
            enum: [attention_desc, attention_asc, name]
            default: attention_desc
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Список участков
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Segment' }
                  total: { type: integer }
  /segments/{segment_id}:
    get:
      tags: [segments]
      summary: Карточка участка
      parameters:
        - $ref: '#/components/parameters/SegmentId'
      responses:
        '200':
          description: Участок
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SegmentDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
  /segments/{segment_id}/scenes:
    get:
      tags: [segments]
      summary: Доступные снимки ДЗЗ по участку
      description: |
        Список разновременных сцен. `tile_url_template` подставляется в MapLibre
        как источник растрового слоя. Дата съёмки обязательна к показу в интерфейсе —
        пользователь должен понимать, что данные не в реальном времени.
      parameters:
        - $ref: '#/components/parameters/SegmentId'
      responses:
        '200':
          description: Сцены, от новых к старым
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Scene' }
        '404': { $ref: '#/components/responses/NotFound' }
  /segments/{segment_id}/subscription:
    put:
      tags: [segments]
      summary: Подписаться на участок
      description: Механика повторного участия — уведомление при появлении нового снимка.
      parameters:
        - $ref: '#/components/parameters/SegmentId'
      responses:
        '204': { description: Подписка оформлена }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [segments]
      summary: Отписаться от участка
      parameters:
        - $ref: '#/components/parameters/SegmentId'
      responses:
        '204': { description: Подписка снята }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ==================== S3: разметка ====================
...
  /moderation/segments:
    get:
      tags: [moderation]
      summary: Очередь модерации разметки
      description: Только роль `oopt_staff`, только в границах своей территории.
      responses:
        '200':
          description: Участки, достигшие консенсуса
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ModerationItem' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /moderation/segments/{segment_id}:
    post:
      tags: [moderation]
      summary: Решение по участку
      description: |
        `approve` переводит участок в статус `work` и повышает репутацию
        разметчиков. `reject` снижает фактор `C`, репутацию и убирает участок
        из очереди. Решение пишется в неизменяемый аудит-лог.
      parameters:
        - $ref: '#/components/parameters/SegmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approve, reject] }
                comment:  { type: string, maxLength: 500 }
      responses:
        '200':
          description: Решение применено
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Segment' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /reports:
    post:
      tags: [reports]
      summary: Создать черновик отчёта (фото «до»)
      description: |
        Шаг 1 сценария S6. Фото «до» обязано быть сделано **до начала уборки** —
        сервер проверяет `taken_at` относительно времени начала акции.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_id, photo_before]
              properties:
                event_id:     { type: string, format: uuid }
                photo_before: { $ref: '#/components/schemas/PhotoRef' }
      responses:
        '201':
          description: Черновик создан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FieldReport' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422':
          description: Нет геометки, времени съёмки или допуск не оформлен
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /reports/{report_id}:
    get:
      tags: [reports]
      summary: Карточка отчёта
      description: Доступна автору отчёта и сотрудникам ООПТ.
      parameters:
        - name: report_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Отчёт
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FieldReport' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [reports]
      summary: Дополнить и отправить отчёт (фото «после»)
      description: Шаг 2 сценария S6. Отправка переводит отчёт в статус `pending`.
      parameters:
        - name: report_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                photo_after: { $ref: '#/components/schemas/PhotoRef' }
                volume_kg:   { type: number, minimum: 0, example: 145 }
                comment:     { type: string, maxLength: 1000 }
                submit:      { type: boolean, default: false, description: true — отправить на модерацию }
      responses:
        '200':
          description: Отчёт обновлён
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FieldReport' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /media/upload:
    post:
      tags: [media]
      summary: Загрузить фото
      description: |
        Возвращает ключ в объектном хранилище. Геометка и время берутся из EXIF,
        а если их нет — принимаются явными полями формы (офлайн-режим PWA).

        Все загруженные фото проходят модерацию до публикации. Для пользователей
        14–17 точная геолокация не публикуется — в публичных слоях она
        огрубляется до участка.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:     { type: string, format: binary }
                lat:      { type: number, description: Если EXIF отсутствует }
                lon:      { type: number }
                taken_at: { type: string, format: date-time }
      responses:
        '201':
          description: Файл загружен
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PhotoRef' }
        '413':
          description: Файл слишком большой (лимит 15 МБ)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ==================== Экспорт ====================
  /export/segments:
    get:
      tags: [export]
      summary: Выгрузка участков для отчётности ООПТ
      parameters:
        - name: format
          in: query
          required: true
          schema: { type: string, enum: [geojson, csv] }
      responses:
        '200':
          description: Файл выгрузки
          content:
            application/geo+json: { schema: { type: object } }
            text/csv:             { schema: { type: string } }
        '403': { $ref: '#/components/responses/Forbidden' }

# ============================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SegmentId:
      name: segment_id
      in: path
      required: true
      schema: { type: string, example: ustie }
...
  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Роль не позволяет выполнить операцию
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message_key]
      properties:
        code:
          type: string
          example: consent_required
          description: Машинный код ошибки
        message_key:
          type: string
          example: err.consent_required
          description: Ключ локализации для i18n.js — текст берётся из словаря фронта
        details:
          type: object
          additionalProperties: true
...
    SegmentStatus:
      type: string
      enum: [problem, work, clean, watch]
      description: |
        * `problem` — выявлена проблема
        * `work` — подтверждён инспектором, в работе
        * `clean` — уборка подтверждена
        * `watch` — под наблюдением
...
    AttentionFactors:
      type: object
      description: |
        Факторы индекса внимания. Индекс = 0.40·S + 0.30·C + 0.20·T + 0.10·A.
        Вычисляется сервером, от клиента не принимается.
      required: [s, c, t, a]
      properties:
        s: { type: number, format: float, minimum: 0, maximum: 1, description: Сигнал ДЗЗ }
        c: { type: number, format: float, minimum: 0, maximum: 1, description: Консенсус людей }
        t: { type: number, format: float, minimum: 0, maximum: 1, description: Давность проверки }
        a: { type: number, format: float, minimum: 0, maximum: 1, description: Доступность участка }
    Segment:
      type: object
      required: [id, name_key, status, attention_index, factors]
      properties:
        id:       { type: string, example: ustie }
        name_key:
          type: string
          example: seg.ustie.name
          description: Ключ локализации — фронт переводит через i18n.js
        oopt_id:  { type: string, example: sinie-ozera }
        length_km: { type: number, example: 2.4 }
        status:   { $ref: '#/components/schemas/SegmentStatus' }
        attention_index:
          type: integer
          minimum: 0
          maximum: 100
          example: 77
          description: Готовое к показу значение 0–100
        factors:  { $ref: '#/components/schemas/AttentionFactors' }
        votes:          { type: integer, example: 3 }
        votes_required: { type: integer, example: 3 }
        verified:       { type: boolean }
        last_verified_at: { type: string, format: date-time, nullable: true }
        is_subscribed:  { type: boolean, description: Подписан ли текущий пользователь }
        is_growing:
          type: boolean
          description: |
            Аномалия растёт от снимка к снимку. Маленькая растущая свалка
            требует внимания раньше, чем большая стабильная — показывайте
            это заметным бейджем.
        growth_rate:
          type: number
          format: float
          nullable: true
          description: Прирост площади аномалии между двумя последними сценами, доля
          example: 0.12
    SegmentDetail:
      allOf:
        - $ref: '#/components/schemas/Segment'
        - type: object
          properties:
            geometry:
              type: object
              description: GeoJSON LineString или Polygon
            history:
              type: array
              items:
                type: object
                required: [at, event_key]
                properties:
                  at:        { type: string, format: date-time }
                  event_key: { type: string, example: history.cleaned }
                  vars:      { type: object, additionalProperties: true }
    Scene:
      type: object
      required: [id, captured_at, source, tile_url_template]
      properties:
        id:          { type: string }
        captured_at: { type: string, format: date }
        source:
          type: string
          enum: [resurs-p, kanopus-v, sentinel-2, uav]
          description: Источник данных — обязателен к показу в интерфейсе
        resolution_m: { type: number, example: 1.0 }
        cloud_cover:  { type: number, minimum: 0, maximum: 1, example: 0.05 }
        tile_url_template:
          type: string
          example: "https://tiles.kosmobereg.ru/scenes/{scene_id}/{z}/{x}/{y}.png"
...
    Mark:
      type: object
      required: [x, y]
      description: Отметка в нормализованных долях тайла [0..1]
      properties:
        x: { type: number, minimum: 0, maximum: 1 }
        y: { type: number, minimum: 0, maximum: 1 }
        r: { type: number, minimum: 0, maximum: 1, description: Радиус зоны, если выделена область }
...
    ModerationItem:
      type: object
      required: [segment, annotations_count]
      properties:
        segment:           { $ref: '#/components/schemas/Segment' }
        annotations_count: { type: integer }
        agreement:
          type: number
          format: float
          description: Доля согласия разметчиков, 0..1
        marks_heatmap:
          type: array
          description: Агрегированные отметки для показа инспектору
          items: { $ref: '#/components/schemas/Mark' }
        queued_at: { type: string, format: date-time }
        annotators_accuracy:
          type: number
          format: float
          nullable: true
          description: |
            Средняя измеренная точность разметчиков участка. Ради неё и
            делались контрольные задания: инспектор видит не «трое что-то
            отметили», а «трое с точностью 87%».
...
    PhotoRef:
      type: object
      required: [key]
      properties:
        key:      { type: string, example: "reports/2026/05/a1b2c3.jpg" }
        url:      { type: string, format: uri, description: Подписанная ссылка на чтение }
        lat:      { type: number, nullable: true }
        lon:      { type: number, nullable: true }
        taken_at: { type: string, format: date-time, nullable: true }
        has_geotag: { type: boolean }
    FieldReport:
      type: object
      required: [id, segment_id, status]
      properties:
        id:         { type: string, format: uuid }
        event_id:   { type: string, format: uuid }
        segment_id: { type: string }
        segment_name_key: { type: string }
        author_name: { type: string, description: Для 14–17 — только имя и первая буква фамилии }
        photo_before: { $ref: '#/components/schemas/PhotoRef' }
        photo_after:  { $ref: '#/components/schemas/PhotoRef' }
        volume_kg:  { type: number, nullable: true }
        comment:    { type: string, nullable: true }
        status:     { type: string, enum: [draft, pending, approved, rejected] }
        submitted_at: { type: string, format: date-time, nullable: true }
        moderated_at: { type: string, format: date-time, nullable: true }

    # ---------- Пакетная модерация и сводка ----------
...